Added city and state to listing gridview in profile fragment. Updated profile styling. Moved logout button from toolbar to settings activity.
