+++ app/src/debug/res/drawable-anydpi/ic_logout.xml
@@ -3,7 +3,7 @@
     android:height="24dp"
     android:viewportWidth="24"
     android:viewportHeight="24"
-    android:tint="@android:color/white"
+    android:tint="?attr/colorPrimary"
     android:alpha="0.6"
     android:autoMirrored="true">
   <path
