+++ app/src/main/res/layout/activity_settings.xml
@@ -49,6 +49,16 @@
         app:layout_constraintStart_toEndOf="@+id/editProfilePicture"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <ImageButton
+        android:id="@+id/logout_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:src="@drawable/ic_logout"
+        android:background="#00000000"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:layout_margin="20dp"/>
+
     <EditText
         android:id="@+id/etEditBio"
         android:layout_width="0dp"
